Add if-else & and-or.
This commit also makes everything a statement. This is something
that I am unsure if I'd like in the final langauge spec, but adds
some interesting behavior. Basically: statements have a return val
that is equal to the last evaluated statement in their body.

For example:

print if (true;) "dog";; // prints dog

The semicolons are sortof insane, but you need them. I'm torn
about this. For now I'll move this over to another branch and keep
things semi normal.